# job-worker-service/internal/entity/job.go
package entity

import (
	"encoding/json"
	"time"

	"github.com/google/uuid"
)

type JobStatus string

const (
	StatusPending    JobStatus = "pending"
	StatusProcessing JobStatus = "processing"
	StatusDone       JobStatus = "done"
	StatusError      JobStatus = "error"
)

type Job struct {
	ID        uuid.UUID       `json:"id"`
	Type      string          `json:"type"`
	Status    JobStatus       `json:"status"`
	Input     json.RawMessage `json:"input"`
	Output    json.RawMessage `json:"output,omitempty"`
	Error     *string         `json:"error,omitempty"`
	CreatedAt time.Time       `json:"created_at"`
	UpdatedAt time.Time       `json:"updated_at"`
	Priority  int             `json:"priority" db:"priority"`
}