# job-worker-service/internal/transport/http/response.go
package httptransport

import (
	"encoding/json"
	"net/http"
)

type apiError struct {
	Message string `json:"message"`
}

func writeJSON(w http.ResponseWriter, code int, v any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(code)
	_ = json.NewEncoder(w).Encode(v)
}

func writeErr(w http.ResponseWriter, code int, msg string) {
	writeJSON(w, code, apiError{Message: msg})
}